introduce helpers to create releases
The ver file contains the current application version.
When Automatix is build with the Makefile, the version string is read
and the Version variable in Automatix's main.go file is set to it.

The release make target creates a tar.xz archive that can be uploaded to
github as release and prints further introductions to finishing the
release.

Automatix now also prints it's version number on every start.